Save pug-cli as an npm package dependency
This will be used to compile the pug files into the HTML files
that will later be deployed to GitHub Pages.

Although npm reports high severity remote code execution vulnerabilities
associated with the pretty option, this package is only really meant
to be used in GitHub Actions prior to deployment to GitHub Pages
so the risk should be acceptable for now.