css: adjust header element sizes and paddings
Completely remove the header's padding
so that elements can move closer to the border.
Increase the padding and font size of the main site link
while also adjusting the padding of the other links.

The navigation links look nicer when closer to the border like this.
The main site link will still have quite a bit of space around it.